+++ src/encodeBase64Url.ts
@@ -1,4 +1,4 @@
-import { toB64 } from './toB64';
+import { encodeBase64 } from './encodeBase64';
 
 /**
  * Converts a Uint8Array to a URL-safe Base64 string.
@@ -10,6 +10,16 @@ import { toB64 } from './toB64';
  * @param {Uint8Array} u8a - The Uint8Array to convert
  * @returns {string} A URL-safe Base64 encoded string
  */
+export const encodeBase64Url = (u8a: Uint8Array) => {
+  return encodeBase64(u8a)
+    .replace(/\+/g, '-')
+    .replace(/\//g, '_')
+    .replace(/=+$/, '');
+};
+
+/**
+ * @deprecated ⚠️ DEPRECATED: Use encodeBase64Url instead ⚠️
+ */
 export const toB64U = (u8a: Uint8Array) => {
-  return toB64(u8a).replace(/\+/g, '-').replace(/\//g, '_').replace(/=+$/, '');
+  return encodeBase64Url(u8a);
 };