+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.11] - 2025-06-11
+
+### Changed
+
+- Renamed `fromHex` to `decodeHex`
+- Renamed `toHex` to `encodeHex`
+- Renamed `fromB64` to `decodeBase64`
+- Renamed `toB64` to `encodeBase64`
+- Renamed `fromB64U` to `decodeBase64Url`
+- Renamed `toB64U` to `encodeBase64Url`
+
+The old function names are now deprecated and will be removed in a future version.
+
 ## [1.0.10] - 2025-06-10
 
 ### Added